refactor: move insert_item_element into submodule with version-gated estimated costs
Extract insert_item_element dispatcher and v0/v1 implementations into
grovedb/src/batch/insert_item_element/{mod,v0,v1}.rs submodule structure.

Add version-gated existence check cost to average_case_cost and
worst_case_cost for InsertOnly operations: in V1+, InsertOnly performs
a merk.exists() seek before inserting, so estimated costs must include
the extra seek + loaded bytes to match actual runtime costs.
