+++ grovedb/src/operations/bulk_append_tree.rs
@@ -525,7 +525,7 @@ impl GroveDb {
             drop(tree);
 
             // Create replacement op
-            // Key is restored for downstream (from_ops, execute_ops_on_path)
+            // Key is restored for downstream (from_ops, insert_item_element)
             let replacement = QualifiedGroveDbOp {
                 path: crate::batch::KeyInfoPath::from_known_owned_path(path_vec),
                 key: Some(crate::batch::key_info::KeyInfo::KnownKey(key_bytes)),
