+++ grovedb/src/operations/mmr_tree.rs
@@ -449,7 +449,7 @@ impl GroveDb {
             drop(storage_ctx);
 
             // Create a ReplaceNonMerkTreeRoot — MMR root flows as child hash
-            // Key is restored for downstream (from_ops, execute_ops_on_path)
+            // Key is restored for downstream (from_ops, insert_item_element)
             let replacement = QualifiedGroveDbOp {
                 path: crate::batch::KeyInfoPath::from_known_owned_path(path_vec),
                 key: Some(crate::batch::key_info::KeyInfo::KnownKey(key_bytes)),
